% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{gp_conditional}
\alias{gp_conditional}
\title{Reconstruct a Gaussian Process conditional on some data.}
\usage{
gp_conditional(theta, acv.model = NULL, dat = NULL, t.star = NULL,
  PDcheck = FALSE)
}
\arguments{
\item{theta}{(vector) parameters for covariance function
the first element is the mean value mu}

\item{acv.model}{(name) name of the function to compute ACV(tau|theta)}

\item{dat}{(data frame) an N * 3 data frame/array, 3 columns}

\item{t.star}{(vector) times at which to compute the simulation(s).}

\item{PDcheck}{(logical) use Matrix::nearPD to coerse the matrix}
}
\value{
A list containing:
  \item{t}{prediction times (same as \code{t.star}.)}
  \item{y}{mean of GP model at times \code{t}.}
  \item{dy}{standard deviation of GP model at times \code{t}.}
  \item{cov}{The full \code{m*m} covariance matrix.}
}
\description{
\code{gp_conditional} returns mean and covariance of a GP given data.
}
\section{Notes}{


Compute the expectation of the GP with covariance matrix specified by
(hyper-)parameters 'theta' at times \code{t.star} based on observations
\code{y} at times \code{t} (in the \code{dat} input).
Uses eqn 2.23 of Rasmussen & Williams (2006).

This may be used for interpolation (between available time points) or
prediction/extrapolation (beyond the range of available time points).
I consider both of these as aspects of `reconstruction'.

In practice we compute \eqn{E[y(t.star)]} given \code{y(t.obs)} and
\code{theta}. This is the mean (as a function of time) of the GP proceess
(with specified ACV) that is conditional on the givendata. We compute not
just the mean but the full covariance matrix for times \code{t.star}:
\code{C[i,j] = ACV(t.star[i], t.star[j])}. This may be a large matrix. Also
returns \code{dy = sqrt(diag(cov))}.
}

\seealso{
\code{\link{gp_sim}}, \code{\link{rmvnorm}}
}
