% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{gp_lagMatrix}
\alias{gp_lagMatrix}
\title{Compute a matrix of differences given two vectors.}
\usage{
gp_lagMatrix(x, y)
}
\arguments{
\item{x}{vector 1}

\item{y}{vector 2 (default to vector 1 if not specified)}
}
\value{
\code{N*M} array of differences, \code{result[i,j] = x[i] - y[j]}
}
\description{
\code{gp_lagMatrix} returns a matrix of differences between two vectors.
}
\details{
Given two vectors - \code{x} (length \code{M}) and \code{y} (length \code{N}) - as
input, return the \code{N*M} matrix of differences \code{result[i,j] = x[i] - y[j]}.
}
\section{Notes}{

Note that in the special case that \code{x=y} we have a square symmetric
matrix: \code{result[i,j] = result[j,i]}. In the even more special case that
the two vectors are evenly spaced (\code{x[i] = y[i] = i * delta + const})
then we have a circulant matrix; the \code{j}th column \code{result[,j]} is
the \code{(j-1)}th cyclic permutation of the first column. This matrix is
symmetric, Toeplitz and circulant.
}

\examples{
result <- gp_lagMatrix(c(1,2,3), c(2,3,4,5,6))
print(result)

}
