% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{gp_logLikelihood}
\alias{gp_logLikelihood}
\title{Compute log likelihood function for Gaussian Process model.}
\usage{
gp_logLikelihood(theta, acv.model = NULL, tau = NULL, dat = NULL,
  PDcheck = TRUE, chatter = 0)
}
\arguments{
\item{theta}{(vector) parameters for covariance function
the first element is the mean value mu}

\item{acv.model}{(name) name of the function to compute ACV(tau | theta)}

\item{tau}{(matrix) N*N matrix of lags at which to compute ACF}

\item{dat}{(matrix) an N * 3 matrix of data: 3 columns}

\item{PDcheck}{(logical) use Matrix::nearPD to coerse the matrix}

\item{chatter}{(integer) higher values give more run-time feedback}
}
\value{
scalar value of log[likelihood(theta)]
}
\description{
\code{gp_logLikelihood} returns the log likelihood for a GP model.
}
\section{Notes}{

 Compute the log likelihood for Gaussian Process model with parameters theta
 given data \eqn{\{t, y, dy\}} and an (optional) N*N matrix of lags, tau. See
 algorithm 2.1 of Rasmussen & Williams (2006). The input data matrix \code{dat}
 should contain three columns: \code{t}, \code{y}, \code{dy}. \code{t[i]} and
 \code{y[i]} give the times and the measured values at those times. \code{dy}
 gives the 'error' on the measurements \code{y}, assumed to be independent
 Gaussian errors wih standard deviation \code{dy}. If \code{dy} is not
 present we assumine \code{dy[i] = 0} for all \code{i}. The columns \code{t},
 \code{y}, and \code{dy} are all \code{n}-element vectors.
}

\seealso{
\code{\link{gp_logPosterior}}
}
