% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{gp_logPosterior}
\alias{gp_logPosterior}
\title{Compute a posterior density given likelihood and prior.}
\usage{
gp_logPosterior(theta, acv.model = NULL, tau = NULL, dat = NULL,
  PDcheck = TRUE, chatter = 0, logPrior = NULL)
}
\arguments{
\item{theta}{(vector) parameters for covariance function
the first element is the mean value mu}

\item{acv.model}{(name) name of the function to compute ACV(tau|theta)}

\item{tau}{(matrix) N*N array of lags at which to compute ACF}

\item{dat}{(data frame) an N * 3 data frame/array, 3 columns}

\item{PDcheck}{(logical) use Matrix::nearPD to coerse the matrix}

\item{chatter}{(integer) higher values give more run-time feedback}

\item{logPrior}{(name) Name of the function returning the log Prior density.}
}
\value{
log posterior density (scalar)
}
\description{
\code{gp_logPosterior} compute posterior density for GP model.
}
\details{
Calculate posterior density for a Gaussian Process (GP) model given some data
(day) and functions for the AutoCovariance (ACV) of the GP, the log
likelihood, and the log prior (both log densities).

Simply computes log(posterior) = log(likelihood) + log(prior), where
log(likelihood) is a function of the data and a ACV (and its parameters), and
log(prior) is a function of the ACV parameters only.
}
\seealso{
\code{\link{gp_logLikelihood}}
}
