% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_snake}
\alias{plot_snake}
\title{Make a 'snake' plot of a Gaussian Process model}
\usage{
plot_snake(dat, sigma = 1, add = FALSE, col.line = "red",
  col.fill = NULL, col.border = NULL, ...)
}
\arguments{
\item{dat}{(data frame) data for time series.}

\item{sigma}{(float) Plot the +/-sigma band (>0).}

\item{add}{(logical) if TRUE then start a new plot}

\item{col.line}{(colour) Colour of the mean line.}

\item{col.fill}{(colour) Colour of the +/-sigma band.}

\item{col.border}{(colour) Colour of +/-sigma limits.}

\item{...}{(anything) any other graphical keywords to be passed to
\code{plot(...)}}
}
\value{
None.
}
\description{
\code{plot_snake} plots the mean and N-sigma band for a GP.
}
\details{
Produces a 'snake' plot. This is a plot showing the mean of a Gaussian
 process against time (as a thick line) and the variance of the process
 illustrated as a band. The lines goes through mu(t) and the band covers
 mu(t) +/- sigma*dy(t) where dy(t) is the standard deviation - this is
 actually sqrt(diag(covariance_matrix)). Sigma is a constant (>0) to allow
 one to plot e.g. the +/-2 std.dev band.
}
\seealso{
\code{\link{plot_ts}}
}
