% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_functions.R
\name{ts_load}
\alias{ts_load}
\title{Load time series data into a standard format array.}
\usage{
ts_load(filename = NA, header = FALSE, csv = FALSE, cols = NA)
}
\arguments{
\item{filename}{(string) name of data file}

\item{header}{(logical) if true then line 1 is a header line.}

\item{csv}{(logical) does the file use commas to seperate columns?}

\item{cols}{(vector) list of columns to read}
}
\value{
array with three named columns \code{t, y, dy}.
}
\description{
\code{ts_load} returns a standard-format data array
}
\section{Notes}{

 Load time series data from a multi-column ASCII file. The input data file
 should contain columns for \code{t}, \code{y} and \code{dy}.
 Here, \code{t} is the time variable, \code{y}
 is the variable measured and \code{dy} is its uncertainty (1-sigma error
 bar). If \code{dy} is not provided then it is assumed to be 0 for all data
 points. The output is a data frame of length N with columnes for \code{t},
 \code{y} and \code{dy}.
 If \code{csv} is set to \code{TRUE} then the file will be read as a CSV
 file. If \code{header} is set to \code{TRUE} then the first line of the
 file is assumed to be a header (and is then ignored).
 The optional input \code{cols} can specify which columns contain the
 \code{t}, \code{y} and \code{dy} data, e.g. \code{cols = c(2,4,5)}.
}

